chealili
Access will not be granted without a complete description.  If the data is being used for graduate level research, include the name of the principal researcher and who is funding the research.The email task feature under Actions works the same as email triggers, but with several notable improvements. Due to these improvements, we recommend using the Email Task instead.Send an email to a customer service rep and their manager every time a customer submits a response to a CSAT survey. This email includes a report of what the customer said about their support interact